refactor: 검색페이지내 탐색 섹션과 검색 결과 섹션 컴포넌트 분리
