+++ client/src/components/search/SearchResult/SearchResult.module.css
@@ -1,0 +1,8 @@
+.container h2 {
+  font-size: 1.5rem;
+  margin-top: 0;
+}
+
+.container h3 {
+  font-size: 1rem;
+}
