+++ client/src/pages/search/search.module.css
@@ -4,62 +4,7 @@
   color: white;
 }
 
-.search {
-  display: flex;
-  flex-direction: column;
-}
-
 .title_input {
   font-size: 0.75rem;
   margin-bottom: 70px;
-}
-
-.hashtag_input {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: 8px;
-  font-size: 0.75rem;
-  margin-bottom: 50px;
-}
-
-.search_history {
-  display: flex;
-  flex-direction: column;
-}
-
-.container h2 {
-  font-size: 1.5rem;
-  margin-top: 0;
-}
-
-.container h3 {
-  font-size: 1rem;
-}
-
-.head {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-weight: 600;
-}
-
-.search_history button {
-  font-size: 0.875rem;
-  padding: 0;
-}
-
-.search_history button:hover {
-  color: rgba(200, 200, 200);
-}
-
-.content li {
-  display: flex;
-  justify-content: space-between;
-  font-size: 0.875rem;
-  margin-bottom: 10px;
-}
-
-.content span:hover {
-  color: rgba(200, 200, 200);
 }
